requirements.txt
streamlit
langchain 
wikipedia 
langchain-community 
langchain-openai
python-dotenv
pydantic
duckduckgo-search
langchain-google-genai
ruff